Fix file-relative error and preview problems on change

+++ docs/_extensions/sverto/refresh.ts
@@ -1,6 +1,13 @@
 // delete the temporary import block files and the temp render metadata
-try {
-  await Deno.remove(".sverto/",  { recursive: true })
-} catch {
 
+console.log("Render all? " + Deno.env.get("QUARTO_PROJECT_RENDER_ALL"))
+
+if (Deno.env.get("QUARTO_PROJECT_RENDER_ALL") == "1") {
+  console.log("Clearing .sverto/")
+  try {
+    await Deno.remove(".sverto/",  { recursive: true })
+  } catch {
+  
+  }
 }
+
